+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.10.0] - 2018-10-06
+### Added
+- Endpoint to delete sample groups.
+- Multi-Axis and Top Taxa fuzz values.
+
 ## [0.9.0] - 2018-05-11
 ### Added
 - Basic Docker configuration.
@@ -19,6 +24,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Authentication endpoints and tests.
 - CircleCI configuration.
 
+### Fixed
+- Updated Celery to 4.2.1 to fix module name change on kombu ([source](https://stackoverflow.com/a/50464774)).
+
 ### Changed
 - Moved API to blueprint.
 
@@ -26,5 +34,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Basic Flask project structure.
 
-[Unreleased]: https://github.com/LongTailBio/metagenscope-server/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/LongTailBio/metagenscope-server/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/LongTailBio/metagenscope-server/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/LongTailBio/metagenscope-server/compare/v0.0.1...v0.9.0